pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "httpx>=0.27.2",
    "mcp>=1.13.1",
    "pydantic-settings>=2.4.0",
    "pytest>=8.4.2",
    "sse-starlette>=2.1.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
backend = "backend.main:main"
